+++ packages/registry/src/index.ts
@@ -44,6 +44,7 @@ import { Uniswap } from '@rabbitholegg/questdk-plugin-uniswap'
 import { Vela } from '@rabbitholegg/questdk-plugin-vela'
 import { WooFi } from '@rabbitholegg/questdk-plugin-woofi'
 import { Zora } from '@rabbitholegg/questdk-plugin-zora'
+import { Foundation } from '@rabbitholegg/questdk-plugin-foundation'
 import { Lens } from '@rabbitholegg/questdk-plugin-lens'
 // ^^^ New Imports Go Here ^^^
 import {
@@ -114,6 +115,7 @@ export const plugins: Record<string, IActionPlugin> = {
   [Superbridge.pluginId]: Superbridge,
   [Neynar.pluginId]: Neynar,
   [Titles.pluginId]: Titles,
+  [Foundation.pluginId]: Foundation,
   [Lens.pluginId]: Lens,
 }
 
@@ -149,6 +151,7 @@ export const getTxSimulation = (
   value: bigint,
   client?: PublicClient,
   account?: Address,
+  creatorAddress?: Address,
 ) => {
   switch (actionType) {
     case ActionType.Mint:
@@ -158,6 +161,7 @@ export const getTxSimulation = (
           value,
           account,
           client,
+          creatorAddress,
         )
       } else {
         throw new PluginActionNotImplementedError()