+++ packages/soundxyz/src/constants.ts
@@ -316,3 +316,7 @@ export const MINT_INFO_LIST_ABI = [
     type: 'function',
   },
 ]
+
+// for referrals
+export const ZORA_DEPLOYER_ADDRESS =
+  '0xe3bBA2A4F8E0F5C32EF5097F988a4d88075C8B48'